// src/app/enum/file-primitives.enum.ts
export enum FilePrimitivesEnum {
  BACKGROUND = 'BACKGROUND',
  LINE = 'LINE',
  BEZIER = '',

  RECTANGLE = 'RECTANGLE',
  FILLED_RECTANGLE = 'FILLED_RECTANGLE',

  ELLIPSE = 'ELLIPSE',
  FILLED_ELLIPSE = 'FILLED_ELLIPSE',

  CIRCLE = 'CIRCLE',
  FILLED_CIRCLE = 'FILLED_CIRCLE',

  TRIANGLE = 'TRIANGLE',
  FILLED_TRIANGLE = 'FILLED TRIANGLE',
}